+++ src/js/util/querySelector.js
@@ -1,0 +1,3 @@
+export const $ = selector => document.querySelector(selector);
+
+export const $$ = selector => document.querySelectorAll(selector);
