+++ src/js/validate.js
@@ -1,0 +1,7 @@
+import { ERROR } from './constants/constants.js';
+
+export const validateInput = input => {
+  if (!input) {
+    throw new Error(ERROR.MESSAGE.EMPTY_INPUT);
+  }
+};
